web_scrapers/news_scraper.py
import requests
from bs4 import BeautifulSoup
from datetime import datetime

def get_real_estate_news():
    """
    Fetches latest real estate and business headlines.
    Returns a list of dicts: {title, url, source, date}
    """
    # In a production environment, you might rotate through 5-10 sources.
    # For this MVP, we target a reliable aggregator structure (simulated for stability).
    
    articles = []
    
    # Target: Generic Business/Property News Structure
    # We will use a User-Agent to act like a real browser
    headers = {
        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.114 Safari/537.36"
    }

    try:
        # Example: Fetching from a standard business feed 
        # (Using a placeholder logic here that you can swap for specific URLs like GhanaWeb or MyJoyOnline)
        # response = requests.get("https://www.ghanaweb.com/GhanaHomePage/realEstate/", headers=headers)
        # soup = BeautifulSoup(response.content, "html.parser")
        
        # SIMULATION FOR MVP STABILITY:
        # Since live scraping usually requires maintaining specific CSS selectors that change weekly,
        # we will simulate the *return format* of a successful scrape so the pipeline works immediately.
        # You can easily swap this block with 'soup.select()' later.
        
        mock_feed = [
            {
                "title": "New Airport City infrastructure project approved in Accra",
                "url": "https://example.com/news/airport-city-expansion",
                "source": "Ghana Business News",
                "published_at": datetime.now().strftime("%Y-%m-%d")
            },
            {
                "title": "Cedi depreciation drives rent prices up in East Legon",
                "url": "https://example.com/news/cedi-rent-hike",
                "source": "Daily Graphic",
                "published_at": datetime.now().strftime("%Y-%m-%d")
            },
            {
                "title": "Road construction begins in Oyibi, sparking investor interest",
                "url": "https://example.com/news/oyibi-road",
                "source": "Joy Online",
                "published_at": datetime.now().strftime("%Y-%m-%d")
            },
             {
                "title": "Flooding concerns rise in Kasoa after heavy rains",
                "url": "https://example.com/news/kasoa-floods",
                "source": "Citi News",
                "published_at": datetime.now().strftime("%Y-%m-%d")
            }
        ]
        articles.extend(mock_feed)

    except Exception as e:
        print(f"⚠️ News Scraper Error: {e}")

    return articles
